Change port to uint16_t

+++ gimli/gimli_client.cc
@@ -3,12 +3,13 @@
 #include "gimli/gimli.grpc.pb.h"
 #include "gimli/gimli.pb.h"
 #include "grpcpp/grpcpp.h"
+#include <cstdint>
 
-ABSL_FLAG(int32_t, port, 8080, "The port where to listen");
+ABSL_FLAG(uint16_t, port, 8080, "The port where to listen");
 
 int main(int argc, char **argv) {
   absl::ParseCommandLine(argc, argv);
-  const int32_t port = absl::GetFlag(FLAGS_port);
+  const uint16_t port = absl::GetFlag(FLAGS_port);
   const std::string address = absl::StrCat("127.0.0.1:", port);
 
   auto channel =
